+++ NEWS.md
@@ -3,6 +3,8 @@ Release 6.0.7 (unreleased)
 
 Bug fixes:
 
+- Ensure profiles are ordered by curvilinear coordinates in serial mode.
+
 - GUI: do not clip enthalpy lower bound to 0 by default.
 
 Release 6.0.6 (December 22 2020)